pyproject.toml
[project]
name = "matriz-atributos"
version = "0.1.0"
description = "Extracción automática de atributos de productos usando Google Gemini"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "google-generativeai>=0.8.0",
    "pandas>=2.0.0",
    "tqdm>=4.66.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
]

[project.optional-dependencies]
dev = [
    "jupyter>=1.0.0",
    "ipykernel>=6.29.0",
]
scraper = [
    "selenium>=4.15.0",
    "playwright>=1.40.0",
]
